pocket_worker/requirements.txt
prometheus_client==0.20.0
Requests==2.32.3
pyyaml
nornir
pythonping